Update Denuncia class Added the nullable = false attribute to the @JoinColumn annotations of both the denunciante and endereco attributes. This means that these foreign key columns cannot be null when persisting a new Denuncia entity.
Added cascade = CascadeType.PERSIST to the denunciante attribute's @ManyToOne annotation, which means that when a new Denuncia is persisted, any new Denunciante entities in the denunciante attribute will also be persisted.

Improved the @OneToOne annotation of the endereco attribute by adding nullable = false to it, which means that an Endereco entity must be present when a new Denuncia is persisted.

No significant changes were made to the existing attributes of the class.